unify normalization of implications in deduction script
