+++ databases/catdat/scripts/utils/functors.ts
@@ -10,16 +10,6 @@ type FunctorMeta = {
 	}
 }
 
-type NormalizedFunctorImplication = {
-	id: string
-	assumptions: Set<string>
-	conclusion: string
-	associated_assumptions: {
-		source: Set<string>
-		target: Set<string>
-	}
-}
-
 /**
  * Returns the list of functors saved in the database along with
  * the satisfied properties of their source and target category.
@@ -76,78 +66,4 @@ export function get_functors(db: Database): FunctorMeta[] {
 			target: parse_json_set<string>(row.target_props),
 		},
 	}))
-}
-
-/**
- * Implications have the form
- * P_1 + ... + P_n ===> Q_1 + ... + Q_m
- * or
- * P_1 + ... + P_n <===> Q_1 + ... + Q_m.
- * This function decomposes them into normalized implications,
- * which have the form
- * P_1 + ... + P_n ===> Q.
- */
-export function get_normalized_functor_implications(
-	db: Database,
-): NormalizedFunctorImplication[] {
-	const all_implications_db = db
-		.prepare<
-			never[],
-			{
-				id: string
-				is_equivalence: 0 | 1
-				assumptions: string
-				source_assumptions: string
-				target_assumptions: string
-				conclusions: string
-			}
-		>(
-			`SELECT
-				id,
-				is_equivalence,
-				assumptions,
-				source_assumptions,
-				target_assumptions,
-				conclusions
-			FROM implications_view i
-			WHERE i.type = 'functor'`,
-		)
-		.all()
-
-	const implications: NormalizedFunctorImplication[] = []
-
-	for (const impl of all_implications_db) {
-		const assumptions = parse_json_set<string>(impl.assumptions)
-		const conclusions = parse_json_set<string>(impl.conclusions)
-		const source_assumptions = parse_json_set<string>(impl.source_assumptions)
-		const target_assumptions = parse_json_set<string>(impl.target_assumptions)
-
-		for (const conclusion of conclusions) {
-			implications.push({
-				id: impl.id,
-				assumptions,
-				conclusion,
-				associated_assumptions: {
-					source: source_assumptions,
-					target: target_assumptions,
-				},
-			})
-		}
-
-		if (impl.is_equivalence) {
-			for (const assumption of assumptions) {
-				implications.push({
-					id: impl.id,
-					assumptions: conclusions,
-					conclusion: assumption,
-					associated_assumptions: {
-						source: source_assumptions,
-						target: target_assumptions,
-					},
-				})
-			}
-		}
-	}
-
-	return implications
 }
